refactor(vocabulary): split 1,279-line flashcard file into focused modules
Extract Context tab, Dictionary tab, shared SoftError widget, and IPA
formatter into separate files to improve navigability and testability.

Closes #394
