+++ lib/features/vocabulary/presentation/vocabulary_flashcard.dart
@@ -1,42 +1,20 @@
-/// Flashcard front/back with Context / Dictionary / Notes tabs.
+/// Flashcard front/back with Context / Dictionary tabs.
 library;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:enjoy_player/core/interaction/enjoy_tappable.dart';
 import 'package:enjoy_player/core/interaction/haptics.dart';
-import 'package:enjoy_player/core/theme/colors.dart';
 import 'package:enjoy_player/core/theme/enjoy_tokens.dart';
-import 'package:enjoy_player/core/theme/lookup_markdown_style.dart';
-import 'package:enjoy_player/core/theme/widgets/enjoy_button.dart';
 import 'package:enjoy_player/core/theme/widgets/enjoy_card.dart';
-import 'package:enjoy_player/features/ai/domain/models/dictionary_result.dart';
-import 'package:enjoy_player/features/auth/application/auth_controller.dart';
-import 'package:enjoy_player/features/auth/domain/auth_state.dart';
-import 'package:enjoy_player/features/auth/presentation/widgets/auth_required_callout.dart';
-import 'package:enjoy_player/features/vocabulary/application/vocabulary_review_media.dart';
-import 'package:enjoy_player/features/vocabulary/application/vocabulary_source_title.dart';
-import 'package:enjoy_player/features/vocabulary/domain/vocabulary_explanation_codec.dart';
 import 'package:enjoy_player/features/vocabulary/domain/vocabulary_models.dart';
 import 'package:enjoy_player/features/vocabulary/presentation/vocabulary_text_style.dart';
-import 'package:enjoy_player/features/vocabulary/presentation/widgets/vocabulary_context_pager.dart';
+import 'package:enjoy_player/features/vocabulary/presentation/widgets/vocabulary_flashcard_context_tab.dart';
+import 'package:enjoy_player/features/vocabulary/presentation/widgets/vocabulary_flashcard_dictionary_tab.dart';
 import 'package:enjoy_player/l10n/app_localizations.dart';
 
-/// Formats IPA with a single leading/trailing slash pair.
-@visibleForTesting
-String formatVocabularyIpa(String raw) {
-  var s = raw.trim();
-  while (s.startsWith('/')) {
-    s = s.substring(1);
-  }
-  while (s.endsWith('/')) {
-    s = s.substring(0, s.length - 1);
-  }
-  s = s.trim();
-  return s.isEmpty ? '' : '/$s/';
-}
+export 'package:enjoy_player/features/vocabulary/presentation/vocabulary_ipa_formatter.dart';
 
 class VocabularyFlashcard extends ConsumerWidget {
   const VocabularyFlashcard({
@@ -371,7 +349,7 @@ class _FlashcardBack extends StatelessWidget {
                   child: TabBarView(
                     children: [
                       _TabBody(
-                        child: _ContextBody(
+                        child: FlashcardContextTab(
                           word: item.word,
                           primaryContext: primaryContext,
                           contextualFetchInFlight: contextualFetchInFlight,
@@ -390,7 +368,7 @@ class _FlashcardBack extends StatelessWidget {
                         ),
                       ),
                       _TabBody(
-                        child: _DictionaryBody(
+                        child: FlashcardDictionaryTab(
                           explanation: item.explanation,
                           fetchInFlight: dictionaryFetchInFlight,
                           error: dictionaryError,
@@ -641,639 +619,6 @@ class _TabBodyState extends State<_TabBody> {
         ),
         child: widget.child,
       ),
-    );
-  }
-}
-
-/// Top-level section heading with icon (Context / Source / Contextual translation).
-class _SectionHeading extends StatelessWidget {
-  const _SectionHeading({required this.icon, required this.label});
-
-  final IconData icon;
-  final String label;
-
-  @override
-  Widget build(BuildContext context) {
-    final t = EnjoyThemeTokens.of(context);
-    final cs = Theme.of(context).colorScheme;
-    final tt = Theme.of(context).textTheme;
-    return Row(
-      children: [
-        Icon(icon, size: 18, color: cs.primary.withValues(alpha: 0.9)),
-        SizedBox(width: t.space8),
-        Expanded(
-          child: Text(
-            label,
-            style: tt.titleSmall?.copyWith(
-              color: cs.onSurface,
-              fontWeight: FontWeight.w600,
-              letterSpacing: -0.1,
-              height: 1.2,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-/// Nested label inside contextual markdown (e.g. part of speech).
-class _DetailLabel extends StatelessWidget {
-  const _DetailLabel(this.text);
-
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    return Text(
-      text,
-      style: Theme.of(context).textTheme.labelSmall?.copyWith(
-        color: cs.onSurfaceVariant.withValues(alpha: 0.75),
-        fontWeight: FontWeight.w600,
-        letterSpacing: 0.35,
-        height: 1.2,
-      ),
-    );
-  }
-}
-
-class _MediaAction extends StatelessWidget {
-  const _MediaAction({
-    required this.icon,
-    required this.label,
-    required this.onPressed,
-  });
-
-  final IconData icon;
-  final String label;
-  final VoidCallback? onPressed;
-
-  @override
-  Widget build(BuildContext context) {
-    final t = EnjoyThemeTokens.of(context);
-    return TextButton.icon(
-      onPressed: onPressed == null
-          ? null
-          : () {
-              Haptics.selection(context);
-              onPressed!();
-            },
-      icon: Icon(icon, size: 18),
-      label: Text(label),
-      style: TextButton.styleFrom(
-        minimumSize: const Size(48, 48),
-        padding: EdgeInsets.symmetric(horizontal: t.space12),
-      ),
-    );
-  }
-}
-
-class _StructuredContextualMarkdown extends StatelessWidget {
-  const _StructuredContextualMarkdown({required this.markdown});
-
-  final String markdown;
-
-  @override
-  Widget build(BuildContext context) {
-    final t = EnjoyThemeTokens.of(context);
-    final theme = Theme.of(context);
-    final cs = theme.colorScheme;
-    final doc = parseContextualMarkdownDocument(markdown);
-    final bodyStyle = buildLookupMarkdownStyleSheet(theme, t).copyWith(
-      p: theme.textTheme.bodyMedium?.copyWith(
-        height: 1.55,
-        color: cs.onSurface.withValues(alpha: 0.92),
-      ),
-      blockSpacing: t.space8,
-      h1: theme.textTheme.bodyMedium,
-      h2: theme.textTheme.bodyMedium,
-      h3: theme.textTheme.bodyMedium,
-      h4: theme.textTheme.bodyMedium,
-      h1Padding: EdgeInsets.zero,
-      h2Padding: EdgeInsets.zero,
-      h3Padding: EdgeInsets.zero,
-      h4Padding: EdgeInsets.zero,
-    );
-    final translationStyle = bodyStyle.copyWith(
-      p: theme.textTheme.bodyLarge?.copyWith(
-        height: 1.5,
-        color: cs.onSurface,
-        fontWeight: FontWeight.w500,
-      ),
-    );
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        if (doc.preamble.isNotEmpty)
-          MarkdownBody(
-            data: doc.preamble,
-            selectable: true,
-            styleSheet: translationStyle,
-          ),
-        for (final section in doc.sections) ...[
-          SizedBox(height: doc.preamble.isNotEmpty ? t.space20 : t.space12),
-          _DetailLabel(section.title),
-          SizedBox(height: t.space8),
-          MarkdownBody(
-            data: section.body,
-            selectable: true,
-            styleSheet: bodyStyle,
-          ),
-        ],
-      ],
-    );
-  }
-}
-
-class _ContextBody extends ConsumerWidget {
-  const _ContextBody({
-    required this.word,
-    required this.primaryContext,
-    required this.contextualFetchInFlight,
-    required this.clipPlayInFlight,
-    required this.contextualError,
-    required this.mediaError,
-    required this.onFetchContextual,
-    required this.onPlayClip,
-    required this.onOpenInPlayer,
-    required this.onShadowReading,
-    required this.contextsCount,
-    required this.activeContextIndex,
-    this.onPreviousContext,
-    this.onNextContext,
-    required this.actionsEnabled,
-  });
-
-  final String word;
-  final VocabularyContext? primaryContext;
-  final bool contextualFetchInFlight;
-  final bool clipPlayInFlight;
-  final String? contextualError;
-  final String? mediaError;
-  final VoidCallback onFetchContextual;
-  final VoidCallback onPlayClip;
-  final VoidCallback onOpenInPlayer;
-  final VoidCallback onShadowReading;
-  final int contextsCount;
-  final int activeContextIndex;
-  final VoidCallback? onPreviousContext;
-  final VoidCallback? onNextContext;
-  final bool actionsEnabled;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context)!;
-    final t = EnjoyThemeTokens.of(context);
-    final theme = Theme.of(context);
-    final cs = theme.colorScheme;
-    final tt = theme.textTheme;
-    final ctx = primaryContext;
-    if (ctx == null || ctx.text.isEmpty) {
-      return Text(l10n.vocabularyNoContextAvailable);
-    }
-
-    final translation = decodeContextualExplanation(ctx.explanation);
-    final canMedia = vocabularyContextSupportsMediaActions(ctx);
-    final locator = ctx.locator;
-    final auth = ref.watch(authCtrlProvider);
-    final signedIn = auth.maybeWhen(
-      data: (s) => s is AuthSignedIn,
-      orElse: () => false,
-    );
-    final titleAsync = ref.watch(vocabularySourceTitleProvider(ctx.sourceId));
-    final sourceTitle =
-        titleAsync.asData?.value ??
-        (titleAsync.isLoading ? null : l10n.vocabularyUnknownSource);
-    final contextBase = tt.bodyLarge?.copyWith(
-      height: 1.55,
-      color: cs.onSurface.withValues(alpha: 0.92),
-    );
-    final contextHighlight = contextBase?.copyWith(
-      fontWeight: FontWeight.w700,
-      color: cs.onSurface,
-      backgroundColor: cs.primary.withValues(alpha: 0.18),
-    );
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        _SectionHeading(
-          icon: Icons.format_quote_rounded,
-          label: l10n.vocabularyContext,
-        ),
-        SizedBox(height: t.space16),
-        DecoratedBox(
-          decoration: BoxDecoration(
-            border: Border(
-              left: BorderSide(
-                width: 3,
-                color: cs.primary.withValues(alpha: 0.45),
-              ),
-            ),
-          ),
-          child: Padding(
-            padding: EdgeInsetsDirectional.only(start: t.space16),
-            child: Text.rich(
-              TextSpan(
-                children: highlightVocabularyWord(
-                  text: ctx.text,
-                  word: word,
-                  base: contextBase ?? const TextStyle(),
-                  highlight: contextHighlight ?? const TextStyle(),
-                ),
-              ),
-            ),
-          ),
-        ),
-        if (contextsCount > 1) ...[
-          SizedBox(height: t.space16),
-          VocabularyContextPager(
-            index: activeContextIndex,
-            total: contextsCount,
-            onPrevious: actionsEnabled ? onPreviousContext : null,
-            onNext: actionsEnabled ? onNextContext : null,
-          ),
-        ],
-        SizedBox(height: t.space32),
-        _SectionHeading(
-          icon: Icons.movie_outlined,
-          label: l10n.vocabularySourceLabel,
-        ),
-        SizedBox(height: t.space16),
-        if (titleAsync.isLoading && sourceTitle == null)
-          Text('…', style: tt.bodyMedium?.copyWith(color: cs.onSurfaceVariant))
-        else
-          Text(
-            sourceTitle ?? l10n.vocabularyUnknownSource,
-            maxLines: 3,
-            overflow: TextOverflow.ellipsis,
-            style: tt.bodyLarge?.copyWith(
-              color: cs.onSurface,
-              height: 1.4,
-              fontWeight: FontWeight.w500,
-              letterSpacing: -0.1,
-            ),
-          ),
-        if (locator != null) ...[
-          SizedBox(height: t.space8),
-          Text(
-            l10n.vocabularyLocatorLabel(
-              (locator.start / 1000).toStringAsFixed(1),
-              (locator.duration / 1000).toStringAsFixed(1),
-            ),
-            style: tt.labelSmall?.copyWith(
-              color: cs.onSurfaceVariant.withValues(alpha: 0.85),
-              letterSpacing: 0.1,
-            ),
-          ),
-        ],
-        if (canMedia) ...[
-          SizedBox(height: t.space8),
-          Wrap(
-            spacing: t.space4,
-            runSpacing: t.space4,
-            children: [
-              _MediaAction(
-                icon: Icons.play_arrow_rounded,
-                label: clipPlayInFlight
-                    ? l10n.vocabularyFetching
-                    : l10n.vocabularyPlaySegment,
-                onPressed: (!actionsEnabled || clipPlayInFlight)
-                    ? null
-                    : onPlayClip,
-              ),
-              _MediaAction(
-                icon: Icons.open_in_new_rounded,
-                label: l10n.vocabularyOpenInPlayer,
-                onPressed: actionsEnabled ? onOpenInPlayer : null,
-              ),
-              _MediaAction(
-                icon: Icons.record_voice_over_rounded,
-                label: l10n.vocabularyEchoReading,
-                onPressed: actionsEnabled ? onShadowReading : null,
-              ),
-            ],
-          ),
-        ] else ...[
-          SizedBox(height: t.space12),
-          Text(
-            l10n.vocabularyMediaUnavailable,
-            style: tt.bodySmall?.copyWith(color: cs.onSurfaceVariant),
-          ),
-        ],
-        if (mediaError != null) ...[
-          SizedBox(height: t.space8),
-          _SoftError(message: l10n.vocabularyMediaPlayFailed),
-        ],
-        SizedBox(height: t.space32),
-        _SectionHeading(
-          icon: Icons.translate_rounded,
-          label: l10n.vocabularyContextualTranslation,
-        ),
-        SizedBox(height: t.space16),
-        if (translation != null)
-          _StructuredContextualMarkdown(markdown: translation.translatedText)
-        else if (!signedIn)
-          const AuthRequiredCallout(
-            surface: AuthRequiredSurface.lookupContextual,
-            compact: true,
-          )
-        else if (contextualFetchInFlight)
-          Padding(
-            padding: EdgeInsets.symmetric(vertical: t.space8),
-            child: Row(
-              children: [
-                SizedBox(
-                  width: 16,
-                  height: 16,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 2,
-                    color: cs.primary.withValues(alpha: 0.85),
-                  ),
-                ),
-                SizedBox(width: t.space12),
-                Text(
-                  l10n.vocabularyFetching,
-                  style: tt.bodyMedium?.copyWith(color: cs.onSurfaceVariant),
-                ),
-              ],
-            ),
-          )
-        else ...[
-          if (contextualError != null)
-            Padding(
-              padding: EdgeInsets.only(bottom: t.space8),
-              child: _SoftError(message: l10n.vocabularyAiFetchFailed),
-            ),
-          Align(
-            alignment: AlignmentDirectional.centerStart,
-            child: EnjoyButton.secondary(
-              onPressed: onFetchContextual,
-              child: Text(l10n.vocabularyFetchContextual),
-            ),
-          ),
-        ],
-      ],
-    );
-  }
-}
-
-class _SoftError extends StatelessWidget {
-  const _SoftError({required this.message});
-
-  final String message;
-
-  @override
-  Widget build(BuildContext context) {
-    final t = EnjoyThemeTokens.of(context);
-    final cs = Theme.of(context).colorScheme;
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        color: cs.errorContainer.withValues(alpha: 0.35),
-        borderRadius: BorderRadius.circular(t.radiusSm),
-      ),
-      child: Padding(
-        padding: EdgeInsets.symmetric(
-          horizontal: t.space12,
-          vertical: t.space8,
-        ),
-        child: Text(
-          message,
-          style: Theme.of(context).textTheme.bodySmall?.copyWith(
-            color: cs.onErrorContainer,
-            height: 1.35,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _DictionaryBody extends ConsumerWidget {
-  const _DictionaryBody({
-    required this.explanation,
-    required this.fetchInFlight,
-    required this.error,
-    required this.onFetch,
-  });
-
-  final String? explanation;
-  final bool fetchInFlight;
-  final String? error;
-  final VoidCallback onFetch;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context)!;
-    final t = EnjoyThemeTokens.of(context);
-    final result = decodeDictionaryExplanation(explanation);
-    final auth = ref.watch(authCtrlProvider);
-    final signedIn = auth.maybeWhen(
-      data: (s) => s is AuthSignedIn,
-      orElse: () => false,
-    );
-
-    if (result != null) {
-      return _DictionaryResultView(result: result);
-    }
-
-    if (!signedIn) {
-      return const AuthRequiredCallout(
-        surface: AuthRequiredSurface.lookupDictionary,
-        compact: true,
-      );
-    }
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Text(l10n.vocabularyDictionaryNotAvailable),
-        if (error != null) ...[
-          SizedBox(height: t.space8),
-          _SoftError(message: l10n.vocabularyAiFetchFailed),
-        ],
-        SizedBox(height: t.space12),
-        EnjoyButton.secondary(
-          onPressed: fetchInFlight ? null : onFetch,
-          child: Text(
-            fetchInFlight
-                ? l10n.vocabularyFetching
-                : l10n.vocabularyFetchDictionary,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _DictionaryResultView extends StatelessWidget {
-  const _DictionaryResultView({required this.result});
-
-  final DictionaryResult result;
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    final t = EnjoyThemeTokens.of(context);
-    final tt = Theme.of(context).textTheme;
-    final scheme = Theme.of(context).colorScheme;
-    final lemmaTrim = result.lemma?.trim();
-    final showLemma =
-        lemmaTrim != null &&
-        lemmaTrim.isNotEmpty &&
-        lemmaTrim != result.word.trim();
-    final ipaFormatted = result.ipa == null
-        ? ''
-        : formatVocabularyIpa(result.ipa!);
-    final hasIpa = ipaFormatted.isNotEmpty;
-
-    if (result.senses.isEmpty) {
-      return Text(l10n.vocabularyDictionaryNotAvailable);
-    }
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        if (hasIpa || showLemma) ...[
-          Wrap(
-            spacing: t.space8,
-            runSpacing: t.space4,
-            crossAxisAlignment: WrapCrossAlignment.center,
-            children: [
-              if (hasIpa)
-                Text(
-                  ipaFormatted,
-                  style: tt.bodySmall?.copyWith(
-                    color: scheme.onSurfaceVariant,
-                    letterSpacing: 0.2,
-                  ),
-                ),
-              if (showLemma)
-                Text(
-                  lemmaTrim,
-                  style: tt.bodySmall?.copyWith(color: scheme.onSurfaceVariant),
-                ),
-            ],
-          ),
-          SizedBox(height: t.space16),
-        ],
-        for (var i = 0; i < result.senses.length; i++) ...[
-          if (i > 0) ...[
-            SizedBox(height: t.space12),
-            Divider(
-              height: 1,
-              color: scheme.outlineVariant.withValues(alpha: 0.18),
-            ),
-            SizedBox(height: t.space12),
-          ],
-          _SenseTile(sense: result.senses[i]),
-        ],
-      ],
-    );
-  }
-}
-
-class _SenseTile extends StatelessWidget {
-  const _SenseTile({required this.sense});
-
-  final DictionarySense sense;
-
-  @override
-  Widget build(BuildContext context) {
-    final tt = Theme.of(context).textTheme;
-    final scheme = Theme.of(context).colorScheme;
-    final t = EnjoyThemeTokens.of(context);
-    final pos = sense.partOfSpeech?.trim();
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-    final translationColor = isDark ? AppColors.brandOnDark : scheme.primary;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        if (pos != null && pos.isNotEmpty) ...[
-          Text(
-            pos,
-            style: tt.labelSmall?.copyWith(
-              color: scheme.primary,
-              fontWeight: FontWeight.w600,
-              letterSpacing: 0.4,
-            ),
-          ),
-          SizedBox(height: t.space4),
-        ],
-        if (sense.definition.trim().isNotEmpty)
-          SelectableText(
-            sense.definition,
-            style: tt.bodyMedium?.copyWith(height: 1.45),
-          ),
-        if (sense.translation != null &&
-            sense.translation!.trim().isNotEmpty) ...[
-          SizedBox(height: t.space4),
-          SelectableText(
-            sense.translation!,
-            style: tt.bodyMedium?.copyWith(
-              fontWeight: FontWeight.w600,
-              color: translationColor,
-              height: 1.4,
-            ),
-          ),
-        ],
-        if (sense.examples != null && sense.examples!.isNotEmpty) ...[
-          SizedBox(height: t.space12),
-          for (final ex in sense.examples!)
-            Padding(
-              padding: EdgeInsets.only(bottom: t.space8),
-              child: Container(
-                decoration: BoxDecoration(
-                  border: Border(
-                    left: BorderSide(
-                      width: 2,
-                      color: scheme.primary.withValues(alpha: 0.4),
-                    ),
-                  ),
-                ),
-                padding: EdgeInsetsDirectional.only(start: t.space12),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    SelectableText(
-                      ex.source,
-                      style: tt.bodySmall?.copyWith(
-                        height: 1.4,
-                        fontStyle: FontStyle.italic,
-                        color: scheme.onSurfaceVariant,
-                      ),
-                    ),
-                    if (ex.target != null && ex.target!.trim().isNotEmpty) ...[
-                      SizedBox(height: t.space4),
-                      SelectableText(
-                        ex.target!,
-                        style: tt.bodySmall?.copyWith(
-                          height: 1.4,
-                          color: scheme.onSurfaceVariant.withValues(
-                            alpha: 0.85,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-            ),
-        ],
-        if (sense.notes != null && sense.notes!.trim().isNotEmpty) ...[
-          SizedBox(height: t.space4),
-          Text(
-            sense.notes!,
-            style: tt.bodySmall?.copyWith(
-              color: scheme.onSurfaceVariant,
-              height: 1.35,
-            ),
-          ),
-        ],
-      ],
     );
   }
 }
